pyproject.toml
[tool.poetry]
name = "lido-assistant"
version = "0.1.0"
description = ""
authors = ["skhomuti <skhomuti@gmail.com>"]
readme = "README.md"
packages = [{include = "lido_assistant"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
openai = {extras = ["datalib"], version = "^0.27.2"}
discord-py = "^2.2.2"
tiktoken = "^0.3.2"
markdown2 = "^2.4.8"
transformers = "^4.27.3"
sentencepiece = "^0.1.97"
torch = "^2.0.0"
tqdm = "^4.65.0"
nltk = "^3.8.1"
langchain = "^0.0.129"
chromadb = "^0.3.21"
unstructured = "^0.5.7"
beautifulsoup4 = "^4.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
